Implement permission handling and refactor CountInPlayer
- Add runtime permission requests for microphone and Bluetooth access in MainActivity and PerformScreen.
- Refactor CountInPlayer to use ToneGenerator instead of SoundPool for click sounds.
- Update LrcLibApi to change duration type from Int to Double for better accuracy.
- Adjust LyricsSearchService to convert duration from Double to Int for compatibility.

This update enhances user experience by ensuring necessary permissions are requested at startup and improves audio handling during performance.

+++ app/src/main/java/com/lyricprompter/data/remote/lyrics/LrcLibApi.kt
@@ -24,7 +24,7 @@ data class LrcLibSearchResult(
     val trackName: String,
     val artistName: String,
     val albumName: String?,
-    val duration: Int?
+    val duration: Double?
 )
 
 data class LrcLibLyrics(
